requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.21.0
gensim>=4.0.0
